Move Spar options to the spar package
