+++ libs/wire-api/src/Wire/API/User/Saml.hs
@@ -38,13 +38,9 @@ import Data.Time
 import GHC.TypeLits (KnownSymbol, symbolVal)
 import GHC.Types (Symbol)
 import Imports
-import SAML2.Util (parseURI', renderURI)
-import SAML2.WebSSO (Assertion, AuthnRequest, ID, IdPId)
-import qualified SAML2.WebSSO as SAML
+import SAML2.WebSSO
 import SAML2.WebSSO.Types.TH (deriveJSONOptions)
-import System.Logger.Extended (LogFormat)
 import URI.ByteString
-import Util.Options
 import Web.Cookie
 import Wire.API.User.Orphans ()
 
@@ -87,37 +83,6 @@ substituteVar var val = substituteVar' ("$" <> var) val . substituteVar' ("%24" 
 substituteVar' :: ST -> ST -> ST -> ST
 substituteVar' var val = ST.intercalate val . ST.splitOn var
 
-type Opts = Opts' DerivedOpts
-
--- FUTUREWORK: Shouldn't these types be in spar, not in wire-api?
-data Opts' a = Opts
-  { saml :: !SAML.Config,
-    brig :: !Endpoint,
-    galley :: !Endpoint,
-    cassandra :: !CassandraOpts,
-    maxttlAuthreq :: !(TTL "authreq"),
-    maxttlAuthresp :: !(TTL "authresp"),
-    -- | The maximum number of SCIM tokens that we will allow teams to have.
-    maxScimTokens :: !Int,
-    -- | The maximum size of rich info. Should be in sync with 'Brig.Types.richInfoLimit'.
-    richInfoLimit :: !Int,
-    -- | Wire/AWS specific; optional; used to discover Cassandra instance
-    -- IPs using describe-instances.
-    discoUrl :: !(Maybe Text),
-    logNetStrings :: !(Maybe (Last Bool)),
-    logFormat :: !(Maybe (Last LogFormat)),
-    -- , optSettings   :: !Settings  -- (nothing yet; see other services for what belongs in here.)
-    derivedOpts :: !a
-  }
-  deriving (Functor, Show, Generic)
-
-instance FromJSON (Opts' (Maybe ()))
-
-data DerivedOpts = DerivedOpts
-  { derivedOptsScimBaseURI :: !URI
-  }
-  deriving (Show, Generic)
-
 -- | (seconds)
 newtype TTL (tablename :: Symbol) = TTL {fromTTL :: Int32}
   deriving (Eq, Ord, Show, Num)
@@ -133,9 +98,6 @@ data TTLError = TTLTooLong String String | TTLNegative String
 
 ttlToNominalDiffTime :: TTL a -> NominalDiffTime
 ttlToNominalDiffTime (TTL i32) = fromIntegral i32
-
-maxttlAuthreqDiffTime :: Opts -> NominalDiffTime
-maxttlAuthreqDiffTime = ttlToNominalDiffTime . maxttlAuthreq
 
 data SsoSettings = SsoSettings
   { defaultSsoCode :: !(Maybe IdPId)
