+++ services/spar/src/Spar/Sem/AReqIDStore/Cassandra.hs
@@ -30,6 +30,7 @@ import Polysemy.Input (Input, input)
 import qualified SAML2.WebSSO as SAML
 import qualified Spar.Data as Data
 import Spar.Data.Instances ()
+import Spar.Options
 import Spar.Sem.AReqIDStore
 import Wire.API.User.Saml
 import Wire.Sem.Now (Now)
