+++ services/spar/test-integration/Test/Spar/APISpec.hs
@@ -75,6 +75,7 @@ import SAML2.WebSSO.Test.Lenses
 import SAML2.WebSSO.Test.MockResponse
 import SAML2.WebSSO.Test.Util
 import qualified Spar.Intra.BrigApp as Intra
+import Spar.Options
 import qualified Spar.Sem.AReqIDStore as AReqIDStore
 import qualified Spar.Sem.BrigAccess as BrigAccess
 import qualified Spar.Sem.IdPConfigStore as IdPEffect
@@ -94,7 +95,6 @@ import Wire.API.User
 import Wire.API.User.Client
 import Wire.API.User.Client.Prekey
 import Wire.API.User.IdentityProvider
-import qualified Wire.API.User.Saml as WireAPI (saml)
 import Wire.API.User.Scim
 
 spec :: SpecWith TestEnv
@@ -151,7 +151,7 @@ specMetadata = do
         mkit mdpath finalizepath = do
           it ("metadata (" <> mdpath <> ")") $ do
             env <- ask
-            let sparHost = env ^. teOpts . to WireAPI.saml . SAML.cfgSPSsoURI . to (cs . SAML.renderURI)
+            let sparHost = env ^. teOpts . to saml . SAML.cfgSPSsoURI . to (cs . SAML.renderURI)
                 fragments =
                   [ "md:SPSSODescriptor",
                     "validUntil",
