+++ services/spar/test-integration/Util/Types.hs
@@ -54,10 +54,10 @@ import Imports
 import SAML2.WebSSO.Types.TH (deriveJSONOptions)
 import Spar.API ()
 import qualified Spar.App as Spar
+import Spar.Options
 import Test.Hspec (pendingWith)
 import Util.Options
 import Wire.API.User.IdentityProvider (WireIdPAPIVersion)
-import Wire.API.User.Saml
 
 type BrigReq = Request -> Request
